backend: fix inconsistent error scheme on http
Prior to this commit the error scheme on the HTTP
interface were inconsistent when using the
connect.ErrorWriter. In this commit we introduce
a new function that allows to write connect.Errors
to the HTTP interface so that it gets serialized
in a consistent format.
