+++ backend/pkg/api/grpc_gateway.go
@@ -11,138 +11,12 @@ package api
 
 import (
 	"context"
-	"errors"
-	"fmt"
-	"io"
 	"net/http"
-	"net/textproto"
 	"strconv"
-	"strings"
 
 	"github.com/grpc-ecosystem/grpc-gateway/v2/runtime"
-	"google.golang.org/genproto/googleapis/rpc/code"
-	spb "google.golang.org/genproto/googleapis/rpc/status"
-	"google.golang.org/grpc/codes"
-	"google.golang.org/grpc/grpclog"
-	"google.golang.org/grpc/status"
 	"google.golang.org/protobuf/proto"
-
-	commonv1alpha1 "github.com/redpanda-data/console/backend/pkg/protogen/redpanda/api/common/v1alpha1"
 )
-
-// NiceHTTPErrorHandler is a clone of grpc-gateway's
-// runtime.DefaultHTTPErrorHandler, with one difference: it uses a modified
-// variant of google.rpc.Status, where code is ENUM instead of int32.
-func NiceHTTPErrorHandler(ctx context.Context, mux *runtime.ServeMux, marshaler runtime.Marshaler, w http.ResponseWriter, r *http.Request, err error) { //nolint:revive // we must follow the function signature demanded by grpc-gateway
-	const fallback = `{"code":"INTERNAL", "message":"failed to marshal error message"}`
-
-	var customStatus *runtime.HTTPStatusError
-	if errors.As(err, &customStatus) {
-		err = customStatus.Err
-	}
-	s := status.Convert(err)
-	pb := StatusToNice(s.Proto())
-
-	w.Header().Del("Trailer")
-	w.Header().Del("Transfer-Encoding")
-
-	contentType := marshaler.ContentType(pb)
-	w.Header().Set("Content-Type", contentType)
-
-	if s.Code() == codes.Unauthenticated {
-		w.Header().Set("WWW-Authenticate", s.Message())
-	}
-
-	buf, merr := marshaler.Marshal(pb)
-	if merr != nil {
-		grpclog.Infof("Failed to marshal error message %q: %v", s, merr)
-		w.WriteHeader(http.StatusInternalServerError)
-		if _, err := io.WriteString(w, fallback); err != nil {
-			grpclog.Infof("Failed to write response: %v", err)
-		}
-		return
-	}
-
-	md, ok := runtime.ServerMetadataFromContext(ctx)
-	if !ok {
-		grpclog.Infof("Failed to extract ServerMetadata from context")
-	}
-
-	handleForwardResponseServerMetadata(w, md)
-
-	// RFC 7230 https://tools.ietf.org/html/rfc7230#section-4.1.2
-	// Unless the request includes a TE header field indicating "trailers"
-	// is acceptable, as described in Section 4.3, a server SHOULD NOT
-	// generate trailer fields that it believes are necessary for the user
-	// agent to receive.
-	doForwardTrailers := requestAcceptsTrailers(r)
-
-	if doForwardTrailers {
-		handleForwardResponseTrailerHeader(w, md)
-		w.Header().Set("Transfer-Encoding", "chunked")
-	}
-
-	st := runtime.HTTPStatusFromCode(s.Code())
-	if customStatus != nil {
-		st = customStatus.HTTPStatus
-	}
-
-	w.WriteHeader(st)
-	if _, err := w.Write(buf); err != nil {
-		grpclog.Infof("Failed to write response: %v", err)
-	}
-
-	if doForwardTrailers {
-		handleForwardResponseTrailer(w, md)
-	}
-}
-
-var defaultOutgoingHeaderMatcher = func(key string) (string, bool) {
-	return fmt.Sprintf("%s%s", runtime.MetadataHeaderPrefix, key), true
-}
-
-func handleForwardResponseServerMetadata(w http.ResponseWriter, md runtime.ServerMetadata) {
-	for k, vs := range md.HeaderMD {
-		if h, ok := defaultOutgoingHeaderMatcher(k); ok {
-			for _, v := range vs {
-				w.Header().Add(h, v)
-			}
-		}
-	}
-}
-
-func requestAcceptsTrailers(req *http.Request) bool {
-	te := req.Header.Get("TE")
-	return strings.Contains(strings.ToLower(te), "trailers")
-}
-
-func handleForwardResponseTrailerHeader(w http.ResponseWriter, md runtime.ServerMetadata) {
-	for k := range md.TrailerMD {
-		tKey := textproto.CanonicalMIMEHeaderKey(fmt.Sprintf("%s%s", runtime.MetadataTrailerPrefix, k))
-		w.Header().Add("Trailer", tKey)
-	}
-}
-
-func handleForwardResponseTrailer(w http.ResponseWriter, md runtime.ServerMetadata) {
-	for k, vs := range md.TrailerMD {
-		tKey := fmt.Sprintf("%s%s", runtime.MetadataTrailerPrefix, k)
-		for _, v := range vs {
-			w.Header().Add(tKey, v)
-		}
-	}
-}
-
-// StatusToNice converts a google.rpc.Status to cloudv1alpha1.ErrorStatus,
-// which is "nicer" variant with Code as Enum.
-func StatusToNice(s *spb.Status) *commonv1alpha1.ErrorStatus {
-	pb := commonv1alpha1.ErrorStatus{
-		Code:    code.Code(s.Code),
-		Message: s.Message,
-		Details: s.Details,
-	}
-
-	return &pb
-}
 
 // GetHTTPResponseModifier returns a ForwardResponseOption that
 // sets a specific http status code, based on a header received from the gRPC
